refactor(notebook-cloud): split hosted route views
