+++ apps/notebook-cloud/viewer/oidc-callback-view.tsx
@@ -1,0 +1,110 @@
+import { useEffect, useState } from "react";
+import { AlertCircle, KeyRound, Loader2, UserRound } from "lucide-react";
+import { useTheme } from "@/hooks/useTheme";
+import { establishCloudAppSessionFromOidcToken } from "./app-session";
+import { completeOidcRedirect } from "./oidc-auth";
+import { applyDocumentTheme, CLOUD_VIEWER_THEME_STORAGE_KEY } from "./theme";
+import type { ViewerStatus } from "./notice-types";
+import type { CloudViewerAuthConfig } from "./cloud-viewer-types";
+
+export function OidcCallbackView({ authConfig }: { authConfig: CloudViewerAuthConfig }) {
+  const { resolvedTheme } = useTheme(CLOUD_VIEWER_THEME_STORAGE_KEY);
+  const [status, setStatus] = useState<ViewerStatus>({
+    kind: "loading",
+    message: "Completing sign-in...",
+  });
+
+  useEffect(() => {
+    applyDocumentTheme(resolvedTheme);
+  }, [resolvedTheme]);
+
+  useEffect(() => {
+    const oidc = authConfig.oidc;
+    if (!oidc) {
+      setStatus({ kind: "error", message: "OIDC sign-in is not configured for this host." });
+      return;
+    }
+
+    const params = new URLSearchParams(window.location.search);
+    if (!params.has("code") || !params.has("state")) {
+      setStatus({ kind: "empty", message: "No sign-in callback is pending." });
+      return;
+    }
+
+    let cancelled = false;
+    void completeOidcRedirect(oidc, {
+      callbackUrl: window.location.href,
+      storage: window.localStorage,
+    })
+      .then(async ({ returnUrl, token }) => {
+        if (cancelled) return;
+        await establishCloudAppSessionFromOidcToken(token).catch((error: unknown) => {
+          console.warn("[notebook-cloud] app session exchange failed", error);
+        });
+        if (cancelled) return;
+        setStatus({ kind: "ready", message: "Signed in. Returning to the notebook..." });
+        window.location.replace(returnUrl);
+      })
+      .catch((error: unknown) => {
+        if (cancelled) return;
+        setStatus({
+          kind: "error",
+          message: error instanceof Error ? error.message : String(error),
+        });
+      });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [authConfig.oidc]);
+
+  const statusTitle =
+    status.kind === "ready"
+      ? "Signed in"
+      : status.kind === "error"
+        ? "Sign-in needs attention"
+        : status.kind === "empty"
+          ? "Nothing to finish"
+          : "Completing sign-in";
+  const statusIcon =
+    status.kind === "error" ? (
+      <AlertCircle aria-hidden="true" />
+    ) : status.kind === "ready" ? (
+      <UserRound aria-hidden="true" />
+    ) : status.kind === "empty" ? (
+      <KeyRound aria-hidden="true" />
+    ) : (
+      <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+    );
+
+  return (
+    <main className="cloud-home">
+      <section className="cloud-home-layout" aria-label="nteract sign-in callback">
+        <div className="cloud-home-copy">
+          <h1>nteract</h1>
+          <span>returning to the notebook</span>
+        </div>
+
+        <section
+          className="cloud-home-panel"
+          data-mode={status.kind}
+          aria-label="Cloud sign-in status"
+        >
+          <div className="cloud-home-status" data-mode={status.kind}>
+            {statusIcon}
+            <div>
+              <h2>{statusTitle}</h2>
+              <p>{status.message}</p>
+            </div>
+          </div>
+
+          {status.kind === "error" || status.kind === "empty" ? (
+            <div className="cloud-home-actions">
+              <a href="/">Back to nteract</a>
+            </div>
+          ) : null}
+        </section>
+      </section>
+    </main>
+  );
+}
